Added new textures. Game pauses after each wave.
Added a path and ground texture for fancy mode. Game now pauses and
waits for player input to continue to the next wave after a wave
finishes. Additionally, player can now click the selected tower button
to deselect it. This allows the player to pan the camera and zoom and
everything without worrying about placing down a tower accidentally.